lib: add first version of `compare_substring`
The implementation seems correct but the semantics are a bit tricky
regarding out of bounds accesses and may be difficult to reimplement
with other backends.

hence, this may change in the future, or another variant of the
function will be provided, but performance will be worse for that
function (computing `length` is not O(1) with all backends,
c.f. `List_of` for example).